Rename project according to new defined name
