--- lib/rails_base.rb
@@ -1,10 +1,0 @@
-require 'rails-base/constants'
-require 'rails-base/cli_actions'
-require 'rails-base/file_manipulation'
-require 'rails-base/git_actions'
-require 'rails-base/rails_actions'
-require 'rails-base/commands'
-require 'rails-base/features'
-require 'rails-base/features/chat'
-require 'rails-base/features/facebook'
-require 'rails-base/features/twilio'
