+++ lib/rs-rails-base/cli_actions.rb
