+++ lib/rs-rails-base/constants.rb
