+++ lib/rs-rails-base/features.rb
