+++ lib/rs-rails-base/features/chat.rb
