+++ lib/rs-rails-base/file_manipulation.rb
