+++ lib/rs-rails-base/git_actions.rb
