+++ lib/rs-rails-base/rails_actions.rb
