+++ lib/rs_rails_base.rb
@@ -1,0 +1,10 @@
+require 'rs-rails-base/constants'
+require 'rs-rails-base/cli_actions'
+require 'rs-rails-base/file_manipulation'
+require 'rs-rails-base/git_actions'
+require 'rs-rails-base/rails_actions'
+require 'rs-rails-base/commands'
+require 'rs-rails-base/features'
+require 'rs-rails-base/features/chat'
+require 'rs-rails-base/features/facebook'
+require 'rs-rails-base/features/twilio'
